fix: fixed the makefile
`nix build` now works without any errors. I made sure that the
derivation was using the actual project directory as the `src` because
that's literally how it should be. I made some changes in the Makefile
so no paths are being hardcoded which could result in issues on NixOS
since packages are stored differently.