feat: placing `bongocat.conf` in `~/.config`
Some users may decide to start Bongocat manually in their compositor
rather than as a SystemD service, by doing this we are placing the
config file generated by Nix in the `~/.config` directory so the user
can manually start Bongocat with the config file that was generated
according to how they configured it in Nix. Just extra convenience.