Monitor selection inside config file
Add a new keyword inside the config file, to select monitor that
wl_surface displays on.

I relied heavely on AI to perform the Wayland implementation, as I have
never worked with the protocol, and I wanted the feature on my own
setup. I expect some review on this code, but somehow it works.